+++ examples/reproducing/csem.py
@@ -146,9 +146,6 @@ ahs = empymod.utils.EMArray(np.nan_to_num(ahs))
 atg = empymod.dipole(**inp3, ab=22)  # " "   Target
 atg = empymod.utils.EMArray(np.nan_to_num(atg))
 
-# Change the phases to unwrapped degrees.
-empymod.EMArray.deg = True
-empymod.EMArray.unwrap = True
 
 ###############################################################################
 # Plot
@@ -160,8 +157,8 @@ plt.subplots_adjust(wspace=.3, hspace=.3)
 # Radial amplitude
 plt.subplot(321)
 plt.title('(a) Radial mode fields')
-plt.plot(x/1000, np.log10(rtg.amp), 'k', label='Model')
-plt.plot(x/1000, np.log10(rhs.amp), 'k-.', label='Half-space response')
+plt.plot(x/1000, np.log10(rtg.amp()), 'k', label='Model')
+plt.plot(x/1000, np.log10(rhs.amp()), 'k-.', label='Half-space response')
 plt.axis([0, 20, -18, -8])
 plt.xlabel('Range (km)')
 plt.ylabel(r'Log$_{10}$(E-field magnitude, V/Am$^2$)')
@@ -170,8 +167,8 @@ plt.legend()
 # Radial phase
 plt.subplot(323)
 plt.title('(b) Radial mode phase')
-plt.plot(x/1000, rtg.pha, 'k')
-plt.plot(x/1000, rhs.pha, 'k-.')
+plt.plot(x/1000, rtg.pha(deg=True), 'k')
+plt.plot(x/1000, rhs.pha(deg=True), 'k-.')
 plt.axis([0, 20, -500, 0])
 plt.xlabel('Range (km)')
 plt.ylabel('Phase (degrees)')
@@ -179,8 +176,8 @@ plt.ylabel('Phase (degrees)')
 # Azimuthal amplitude
 plt.subplot(325)
 plt.title('(c) Azimuthal mode fields')
-plt.plot(x/1000, np.log10(atg.amp), 'k', label='Model')
-plt.plot(x/1000, np.log10(ahs.amp), 'k-.', label='Half-space response')
+plt.plot(x/1000, np.log10(atg.amp()), 'k', label='Model')
+plt.plot(x/1000, np.log10(ahs.amp()), 'k-.', label='Half-space response')
 plt.axis([0, 20, -18, -8])
 plt.xlabel('Range (km)')
 plt.ylabel(r'Log$_{10}$(E-field magnitude, V/Am$^2$)')
@@ -189,8 +186,8 @@ plt.legend()
 # Azimuthal phase
 plt.subplot(322)
 plt.title('(d) Azimuthal mode phase')
-plt.plot(x/1000, atg.pha+180, 'k')
-plt.plot(x/1000, ahs.pha+180, 'k-.')
+plt.plot(x/1000, atg.pha(deg=True)+180, 'k')
+plt.plot(x/1000, ahs.pha(deg=True)+180, 'k-.')
 plt.axis([0, 20, -500, 0])
 plt.xlabel('Range (km)')
 plt.ylabel('Phase (degrees)')